# contracts/walksafe.openapi.json
{
  "components": {
    "schemas": {
...
      "AndroidDebugDepthLogRequest": {
        "additionalProperties": false,
        "properties": {
          "android_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Android Version"
          },
          "app_version_name": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Version Name"
          },
          "device_model": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Model"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AndroidDebugDepthLogEntry"
            },
            "maxItems": 30,
            "minItems": 1,
            "title": "Entries",
            "type": "array"
          },
          "schema_version": {
            "const": "android.depth_debug.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "session_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "session_id",
          "entries"
        ],
        "title": "AndroidDebugDepthLogRequest",
        "type": "object"
      },
...
      "Body_create_frame_capture_android_debug_frame_captures_post": {
        "properties": {
          "image": {
            "format": "binary",
            "title": "Image",
            "type": "string"
          },
          "metadata": {
            "title": "Metadata",
            "type": "string"
          }
        },
        "required": [
          "metadata",
          "image"
        ],
        "title": "Body_create_frame_capture_android_debug_frame_captures_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "WalkSafeAccountGeneration": {
        "description": "Positive account generation bound by the gateway actor assertion.",
        "in": "header",
        "name": "X-WalkSafe-Account-Generation",
        "type": "apiKey"
      },
      "WalkSafeActorAssertion": {
        "description": "Short-lived gateway assertion binding the actor to its authenticated role.",
        "in": "header",
        "name": "X-WalkSafe-Actor-Assertion",
        "type": "apiKey"
      },
      "WalkSafeActorId": {
        "description": "Named actor for secured, rate-limited, or audited operations.",
        "in": "header",
        "name": "X-WalkSafe-Actor-Id",
        "type": "apiKey"
      },
      "WalkSafeAdminAppKind": {
        "description": "Required exact value: ADMIN_ANDROID.",
        "in": "header",
        "name": "X-WalkSafe-App-Kind",
        "type": "apiKey",
        "x-walksafe-required-value": "ADMIN_ANDROID"
      },
      "WalkSafeAdminAudience": {
        "description": "Required exact value: walksafe-admin-api.",
        "in": "header",
        "name": "X-WalkSafe-Audience",
        "type": "apiKey",
        "x-walksafe-required-value": "walksafe-admin-api"
      },
      "WalkSafeAdminBearer": {
        "bearerFormat": "opaque per-device session",
        "description": "Database-backed administrator session; static administrator tokens are rejected.",
        "scheme": "bearer",
        "type": "http"
      },
      "WalkSafeAdminDeviceChallengeId": {
        "description": "Canonical UUID of the fresh one-use device challenge.",
...
  },
  "info": {
    "title": "WalkSafe Assist API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/android/debug/depth-logs": {
      "post": {
        "operationId": "create_depth_log_android_debug_depth_logs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AndroidDebugDepthLogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Depth Log Android Debug Depth Logs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "WalkSafeAdminAppKind": [],
            "WalkSafeAdminAudience": [],
            "WalkSafeAdminBearer": [],
            "WalkSafeAdminDeviceId": [],
            "WalkSafeAdminRole": []
          }
        ],
        "summary": "Create Depth Log",
        "tags": [
          "android-debug"
        ],
        "x-walksafe-actor-identity-required": false,
        "x-walksafe-actor-identity-source": "admin-session",
        "x-walksafe-required-role": "admin"
      }
    },
    "/android/debug/depth-logs/recent": {
      "get": {
        "operationId": "recent_depth_logs_android_debug_depth_logs_recent_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Recent Depth Logs Android Debug Depth Logs Recent Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "WalkSafeAdminAppKind": [],
            "WalkSafeAdminAudience": [],
            "WalkSafeAdminBearer": [],
            "WalkSafeAdminDeviceId": [],
            "WalkSafeAdminRole": []
          }
        ],
        "summary": "Recent Depth Logs",
        "tags": [
          "android-debug"
        ],
        "x-walksafe-actor-identity-required": false,
        "x-walksafe-actor-identity-source": "admin-session",
        "x-walksafe-required-role": "admin"
      }
    },
    "/android/debug/frame-captures": {
      "post": {
        "operationId": "create_frame_capture_android_debug_frame_captures_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_frame_capture_android_debug_frame_captures_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Frame Capture Android Debug Frame Captures Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "WalkSafeAdminAppKind": [],
            "WalkSafeAdminAudience": [],
            "WalkSafeAdminBearer": [],
            "WalkSafeAdminDeviceId": [],
            "WalkSafeAdminRole": []
          }
        ],
        "summary": "Create Frame Capture",
        "tags": [
          "android-debug"
        ],
        "x-walksafe-actor-identity-required": false,
        "x-walksafe-actor-identity-source": "admin-session",
        "x-walksafe-required-role": "admin"
      }
    },
    "/android/debug/frame-captures/recent": {
      "get": {
        "operationId": "recent_frame_captures_android_debug_frame_captures_recent_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Recent Frame Captures Android Debug Frame Captures Recent Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "WalkSafeAdminAppKind": [],
            "WalkSafeAdminAudience": [],
            "WalkSafeAdminBearer": [],
            "WalkSafeAdminDeviceId": [],
            "WalkSafeAdminRole": []
          }
        ],
        "summary": "Recent Frame Captures",
        "tags": [
          "android-debug"
        ],
        "x-walksafe-actor-identity-required": false,
        "x-walksafe-actor-identity-source": "admin-session",
        "x-walksafe-required-role": "admin"
      }
    },
...
  },
...
}